+++ e2e/selective-run-in-band/run-in-band-project-1/jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  displayName: 'run-in-band-project-1',
+  runInBand: true,
+  testEnvironment: 'node',
+};