+++ packages/jest-config/src/normalize.ts
@@ -992,6 +992,7 @@ export default async function normalize(
       case 'resetModules':
       case 'restoreMocks':
       case 'rootDir':
+      case 'runInBand':
       case 'runTestsByPath':
       case 'silent':
       case 'skipFilter':