+++ packages/jest-types/src/Config.ts
@@ -100,6 +100,7 @@ export type DefaultOptions = {
   resetModules: boolean;
   restoreMocks: boolean;
   roots: Array<Path>;
+  runInBand: boolean;
   runTestsByPath: boolean;
   runner: string;
   setupFiles: Array<Path>;
@@ -210,6 +211,7 @@ export type InitialOptions = Partial<{
   restoreMocks: boolean;
   rootDir: Path;
   roots: Array<Path>;
+  runInBand: boolean;
   runner: string;
   runTestsByPath: boolean;
   /**
@@ -325,6 +327,7 @@ export type GlobalConfig = {
   projects: Array<Glob>;
   replname?: string;
   reporters?: Array<string | ReporterConfig>;
+  runInBand: boolean;
   runTestsByPath: boolean;
   rootDir: Path;
   silent?: boolean;
@@ -384,6 +387,7 @@ export type ProjectConfig = {
   restoreMocks: boolean;
   rootDir: Path;
   roots: Array<Path>;
+  runInBand: boolean;
   runner: string;
   setupFiles: Array<Path>;
   setupFilesAfterEnv: Array<Path>;